+++ lib/Frontend_CREDO/LoginPage.dart
@@ -1,0 +1,106 @@
+import 'package:flutter/material.dart';
+import '../network/repository.dart';
+import '../main.dart';
+
+// Application Login Screen
+class LoginPage extends StatefulWidget {
+  @override
+  _LoginPageState createState() => _LoginPageState();
+}
+
+class _LoginPageState extends State<LoginPage> {
+  // login & password field values
+  late String _login;
+  late String _password;
+
+  // Class the handles all interaction with CREDO API's
+  CredoRepository _credoRepository = CredoRepository();
+  TextStyle style = TextStyle(fontFamily: 'Montserrat', fontSize: 20.0);
+
+  @override
+  void initState() {
+    super.initState();
+    //initialise class to load system and device info
+    _credoRepository.init();
+  }
+
+  @override
+  Widget build(BuildContext context) {
+    //Email/username field
+    final emailField = TextField(
+        obscureText: false,
+        style: style,
+        decoration: InputDecoration(
+            contentPadding: EdgeInsets.fromLTRB(20.0, 15.0, 20.0, 15.0),
+            hintText: "Username/Email",
+            border:
+                OutlineInputBorder(borderRadius: BorderRadius.circular(32.0))),
+        onChanged: (value) {
+          _login = value;
+        });
+
+    //Password Field
+    final passwordField = TextField(
+        obscureText: true,
+        style: style,
+        decoration: InputDecoration(
+            contentPadding: EdgeInsets.fromLTRB(20.0, 15.0, 20.0, 15.0),
+            hintText: "Password",
+            border:
+                OutlineInputBorder(borderRadius: BorderRadius.circular(32.0))),
+        onChanged: (value) {
+          _password = value;
+        });
+    //Login Field
+    final loginButton = Material(
+      elevation: 5.0,
+      borderRadius: BorderRadius.circular(30.0),
+      color: Color(0xff01A0C7),
+      child: MaterialButton(
+        minWidth: MediaQuery.of(context).size.width,
+        padding: EdgeInsets.fromLTRB(20.0, 15.0, 20.0, 15.0),
+        onPressed: () {
+          // send login request to endpoint
+          _credoRepository.requestLogin(_login, _password);
+          // navigate to detection screen
+          Navigator.push(
+            context,
+            MaterialPageRoute(builder: (context) => MyHomePage()),
+          );
+        },
+        child: Text("Login",
+            textAlign: TextAlign.center,
+            style: style.copyWith(
+                color: Colors.white, fontWeight: FontWeight.bold)),
+      ),
+    );
+
+    return Scaffold(
+      body: Center(
+        child: Container(
+          color: Colors.white,
+          child: Padding(
+            padding: const EdgeInsets.all(36.0),
+            child: Column(
+              crossAxisAlignment: CrossAxisAlignment.start,
+              mainAxisAlignment: MainAxisAlignment.center,
+              children: <Widget>[
+                SizedBox(height: 45.0),
+                emailField,
+                SizedBox(height: 25.0),
+                passwordField,
+                SizedBox(
+                  height: 35.0,
+                ),
+                loginButton,
+                SizedBox(
+                  height: 15.0,
+                ),
+              ],
+            ),
+          ),
+        ),
+      ),
+    );
+  }
+}
