+++ lib/Hit.dart
@@ -7,27 +7,27 @@ class Hit {
   //@SerializedName("frame_content");
   var frameContent = null; //string?
   var timestamp = null;
-  double latitude = null;
-  double longitude = null;
-  double altitude = null;
-  double accuracy = null; //float
+  double latitude = 0;
+  double longitude = 0;
+  double altitude = 0;
+  double accuracy = 0; //float
   var provider = null; // srting?
-  int width = null;
-  int height = null;
-  int x = null;
-  int y = null;
+  int width = 0;
+  int height = 0;
+  int x = 0;
+  int y = 0;
   //@SerializedName("max");
-  int maxValue = null;
-  double average = null; //float
+  int maxValue = 0;
+  double average = 0; //float
   //@SerializedName("blacks");
-  double blacksPercentage = null; //float
+  double blacksPercentage = 0; //float
   //@SerializedName("black_threshold");
-  int blackThreshold = null;
-  double ax = null; //float
-  double ay = null; //float
-  double az = null; //float
-  double orientation = null; //float
-  int temperature = null;
+  int blackThreshold = 0;
+  double ax = 0; //float
+  double ay = 0; //float
+  double az = 0; //float
+  double orientation = 0; //float
+  int temperature = 0;
 
   // return json represntation of a hit object
   Map<String, dynamic> toJson() => {
