+++ lib/main.dart
@@ -1,12 +1,11 @@
 import 'package:credo_transcript/AllSensorsHelper.dart';
 import 'package:flutter/material.dart';
-import 'FileUtils.dart';
 import 'Frontend_CREDO/AccountsPage.dart';
 import 'Frontend_CREDO/HomePage.dart';
 import 'Frontend_CREDO/DetectorPage.dart';
 import 'Frontend_CREDO/SciencePagePage.dart';
 import 'Frontend_CREDO/HelpPage.dart';
-import 'Frontend_CREDO/AccountsPage.dart';
+import 'network/repository.dart';
 import 'Frontend_CREDO/DetectorSettingsPage.dart';
 import 'Frontend_CREDO/themeSettings.dart';
 
@@ -33,32 +32,21 @@ class CredoHome extends StatelessWidget {
   @override
   Widget build(BuildContext context) {
     return MaterialApp(
-      theme: credoTheme(),
-      title: _title,
-      home: MyHomePage(),
-      routes: {
-        // Routes.detectorStatisticsPage: (BuildContext context) => detector,
-        Routes.accountsPage: (BuildContext context) => AccountsPage(),
-        Routes.detectorSettingsPage: (BuildContext context) =>
-            detectorSettingsPage(),
-      },
-      title: 'Flutter Demo',
-      theme: ThemeData(
-        //body that will build the app when opened
-        primarySwatch: Colors.blue,
-      ),
-      home: LoginPage(
-        title: 'CREDO Login Page',
-      ),
-    );
+        theme: credoTheme(),
+        title: _title,
+        home: MyHomePage(),
+        routes: {
+          // Routes.detectorStatisticsPage: (BuildContext context) => detector,
+          Routes.accountsPage: (BuildContext context) => AccountsPage(),
+          Routes.detectorSettingsPage: (BuildContext context) =>
+              detectorSettingsPage(),
+        });
+
+    //home: LoginPage(
   }
 }
 
 class MyHomePage extends StatefulWidget {
-  MyHomePage({Key? key, required this.title}) : super(key: key);
-
-  final String title;
-
   @override
   _MyHomePageState createState() => _MyHomePageState();
 }
@@ -170,14 +158,6 @@ class _MyHomePageState extends State<MyHomePage> {
                 Navigator.pushNamed(context, Routes.detectorStatisticsPage);
               },
             ),
-            Text(fileContents),
-            RaisedButton(
-              child: Text("Logout"),
-              onPressed: () {
-                _credoRepository.clearPrefs();
-                Navigator.pop(context);
-              },
-            ),
           ],
         ),
       ),
@@ -207,115 +187,6 @@ class _MyHomePageState extends State<MyHomePage> {
         //unselectedItemColor: Colors.black54,
         //backgroundColor: Colors.blueGrey,
         onTap: _onItemTapped,
-      ),
-    );
-  }
-}
-
-// Application Login Screen
-class LoginPage extends StatefulWidget {
-  LoginPage({Key key, this.title}) : super(key: key);
-
-  final String title;
-
-  @override
-  _LoginPageState createState() => _LoginPageState();
-}
-
-class _LoginPageState extends State<LoginPage> {
-  // login & password field values
-  String _login;
-  String _password;
-
-  // Class the handles all interaction with CREDO API's
-  CredoRepository _credoRepository = CredoRepository();
-  TextStyle style = TextStyle(fontFamily: 'Montserrat', fontSize: 20.0);
-
-  @override
-  void initState() {
-    super.initState();
-    //initialise class to load system and device info
-    _credoRepository.init();
-  }
-
-  @override
-  Widget build(BuildContext context) {
-    //Email/username field
-    final emailField = TextField(
-        obscureText: false,
-        style: style,
-        decoration: InputDecoration(
-            contentPadding: EdgeInsets.fromLTRB(20.0, 15.0, 20.0, 15.0),
-            hintText: "Username/Email",
-            border:
-                OutlineInputBorder(borderRadius: BorderRadius.circular(32.0))),
-        onChanged: (value) {
-          _login = value;
-        });
-
-    //Password Field
-    final passwordField = TextField(
-        obscureText: true,
-        style: style,
-        decoration: InputDecoration(
-            contentPadding: EdgeInsets.fromLTRB(20.0, 15.0, 20.0, 15.0),
-            hintText: "Password",
-            border:
-                OutlineInputBorder(borderRadius: BorderRadius.circular(32.0))),
-        onChanged: (value) {
-          _password = value;
-        });
-    //Login Field
-    final loginButton = Material(
-      elevation: 5.0,
-      borderRadius: BorderRadius.circular(30.0),
-      color: Color(0xff01A0C7),
-      child: MaterialButton(
-        minWidth: MediaQuery.of(context).size.width,
-        padding: EdgeInsets.fromLTRB(20.0, 15.0, 20.0, 15.0),
-        onPressed: () {
-          // send login request to endpoint
-          _credoRepository.requestLogin(_login, _password);
-          // navigate to detection screen
-          Navigator.push(
-            context,
-            MaterialPageRoute(
-                builder: (context) =>
-                    MyHomePage(title: 'CREDO Demo Home Page')),
-          );
-        },
-        child: Text("Login",
-            textAlign: TextAlign.center,
-            style: style.copyWith(
-                color: Colors.white, fontWeight: FontWeight.bold)),
-      ),
-    );
-
-    return Scaffold(
-      body: Center(
-        child: Container(
-          color: Colors.white,
-          child: Padding(
-            padding: const EdgeInsets.all(36.0),
-            child: Column(
-              crossAxisAlignment: CrossAxisAlignment.start,
-              mainAxisAlignment: MainAxisAlignment.center,
-              children: <Widget>[
-                SizedBox(height: 45.0),
-                emailField,
-                SizedBox(height: 25.0),
-                passwordField,
-                SizedBox(
-                  height: 35.0,
-                ),
-                loginButton,
-                SizedBox(
-                  height: 15.0,
-                ),
-              ],
-            ),
-          ),
-        ),
       ),
     );
   }
